backend/requirements.txt
# SafeLink Shield Backend Dependencies
# Install with: pip install -r requirements.txt

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic[email]>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
alembic>=1.12.0

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Image Processing & OCR
Pillow>=10.0.0
pytesseract>=0.3.10

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# File Upload Support
python-multipart>=0.0.6

# ML/AI Libraries
transformers>=4.35.0
huggingface-hub>=0.19.0

# OpenAI (optional)
openai>=1.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Local Speech-to-Text (optional, for offline transcription)
# Uncomment if you want local Whisper:
# faster-whisper>=0.9.0
